ExecutableMonitor/requirements.txt
watchdog>=2.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0